Rename throwingTests rule to noForceTryInTests
